refactor main class
